+++ src/commands/prettifyDocumentCommand.ts
@@ -3,7 +3,6 @@ import * as vscode from "vscode";
 import { EdiUtils } from "../utils/ediUtils";
 import * as constants from "../constants";
 import Utils from "../utils/utils";
-import { EdiFormattingOptions } from "../parser/entities";
 
 export class PrettifyDocumentCommand implements ICommandable {
   name: string = constants.commands.prettifyDocumentCommand.name;
@@ -25,13 +24,6 @@ export class PrettifyDocumentCommand implements ICommandable {
       return;
     }
 
-    let formattingOptions: EdiFormattingOptions;
-    if (vscode.workspace.getConfiguration(constants.configuration.ediSupport).get(constants.configuration.formatting.indentSegmentsInLoop)) {
-      formattingOptions = new EdiFormattingOptions(Utils.getStringAsInt(vscodeEditorOptions.tabSize) ?? 2, Utils.getValueAsBoolean(vscodeEditorOptions.insertSpaces, true));
-    } else {
-      formattingOptions = EdiFormattingOptions.TAB_SIZE_0;
-    }
-
     vscode.window.activeTextEditor.edit((builder) => {
       if (!vscode.window.activeTextEditor) {
         return;
@@ -42,7 +34,7 @@ export class PrettifyDocumentCommand implements ICommandable {
           vscode.window.activeTextEditor.document.positionAt(0), 
           vscode.window.activeTextEditor.document.positionAt(documentContent.length)
         ),
-        ediDocument.getFormatString(formattingOptions)
+        ediDocument.getFormatString()
       );
     });
   }