+++ src/decorations/ediDecorations.ts
@@ -1,6 +1,7 @@
 import * as vscode from "vscode";
 import { IDecorationable } from "../interfaces/decorationable";
 import { EdiDecorationsAfterLineProvider } from "./ediDecorationsAfterLineProvider";
+import { EdiDecorationsBeforeLineProvider } from "./ediDecorationsBeforeLineProvider";
 import { EdiDecorationsBeforePositionProvider } from "./ediDecorationsBeforePositionProvider";
 
 export class EdiDecorationsMgr implements IDecorationable {
@@ -9,6 +10,7 @@ export class EdiDecorationsMgr implements IDecorationable {
 
   constructor() {
     this.decorationsProviders.push(new EdiDecorationsAfterLineProvider());
+    this.decorationsProviders.push(new EdiDecorationsBeforeLineProvider());
     // this.decorationsProviders.push(new EdiDecorationsBeforePositionProvider());
   }