+++ src/providers/documentFormattingEdiProvider.ts
@@ -1,8 +1,7 @@
 import * as vscode from "vscode";
 import { IProvidable } from "../interfaces/providable";
-import { EdiFormattingOptions, EdiType } from "../parser/entities";
+import { EdiType } from "../parser/entities";
 import { EdiUtils } from "../utils/ediUtils";
-import * as constants from "../constants";
 
 export class DocumentFormattingEditEdiProvider implements vscode.DocumentFormattingEditProvider, IProvidable {
   async provideDocumentFormattingEdits(document: vscode.TextDocument, options: vscode.FormattingOptions, token: vscode.CancellationToken): Promise<vscode.TextEdit[] | null | undefined> {
@@ -14,20 +13,13 @@ export class DocumentFormattingEditEdiProvider implements vscode.DocumentFormatt
       return;
     }
 
-    let formattingOptions: EdiFormattingOptions;
-    if (vscode.workspace.getConfiguration(constants.configuration.ediSupport).get(constants.configuration.formatting.indentSegmentsInLoop)) {
-      formattingOptions = new EdiFormattingOptions(options.tabSize, options.insertSpaces);
-    } else {
-      formattingOptions = EdiFormattingOptions.TAB_SIZE_0;
-    }
-
     const result: vscode.TextEdit[] = [];
     result.push(new vscode.TextEdit(
       new vscode.Range(
         document.positionAt(0),
         document.positionAt(document.getText().length)
       ),
-      ediDocument.getFormatString(formattingOptions)
+      ediDocument.getFormatString()
     ));
     return result;
   }